+++ ICD.Connect.Telemetry.Crestron/Assets/IFusionAssetDataCollection.cs
@@ -1,6 +1,8 @@
-﻿namespace ICD.Connect.Telemetry.Crestron.Assets
+﻿using System.Collections.Generic;
+
+namespace ICD.Connect.Telemetry.Crestron.Assets
 {
-	public interface IFusionAssetDataCollection
+	public interface IFusionAssetDataCollection : IEnumerable<FusionAssetData>
 	{
 		/// <summary>
 		/// Gets the fusion asset data with the given id.