+++ ICD.Connect.Telemetry.CrestronPro/Assets/FusionStaticAssetAdapter.cs
@@ -1,7 +1,10 @@
-﻿using ICD.Connect.Telemetry.CrestronPro.Devices;
+﻿using ICD.Connect.Protocol.Sigs;
+using ICD.Connect.Telemetry.CrestronPro.Devices;
 using System.Collections.Generic;
+using ICD.Connect.Telemetry.Nodes;
 using eAssetType = ICD.Connect.Telemetry.Crestron.Assets.eAssetType;
 #if SIMPLSHARP
+using Crestron.SimplSharpPro;
 using Crestron.SimplSharpPro.Fusion;
 using ICD.Connect.Telemetry.Crestron.Assets;
 
@@ -82,6 +85,24 @@ namespace ICD.Connect.Telemetry.CrestronPro.Assets
 		public string ReadSerialSig(uint sig)
 		{
 			return m_Asset.FusionGenericAssetSerialsAsset3.StringOutput[sig + FusionRoomAdapter.SIG_OFFSET].StringValue;
+		}
+
+		public IEnumerable<KeyValuePair<SigInfo, eTelemetryIoMask>> GetUserDefinedBooleanSigs()
+		{
+			foreach (BooleanSigData sig in m_Asset.FusionGenericAssetDigitalsAsset1.UserDefinedBooleanSigDetails)
+				yield return new KeyValuePair<SigInfo, eTelemetryIoMask>(new SigInfo(ICD.Connect.Protocol.Sigs.eSigType.Digital, sig.Number, sig.Name, 0), sig.SigIoMask.ToIcd());
+		}
+
+		public IEnumerable<KeyValuePair<SigInfo, eTelemetryIoMask>> GetUserDefinedUShortSigs()
+		{
+			foreach (UShortSigData sig in m_Asset.FusionGenericAssetAnalogsAsset2.UserDefinedUShortSigDetails)
+				yield return new KeyValuePair<SigInfo, eTelemetryIoMask>(new SigInfo(ICD.Connect.Protocol.Sigs.eSigType.Analog, sig.Number, sig.Name, 0), sig.SigIoMask.ToIcd());
+		}
+
+		public IEnumerable<KeyValuePair<SigInfo, eTelemetryIoMask>> GetUserDefinedSerialSigs()
+		{
+			foreach (StringSigData sig in m_Asset.FusionGenericAssetSerialsAsset3.UserDefinedStringSigDetails)
+				yield return new KeyValuePair<SigInfo, eTelemetryIoMask>(new SigInfo(ICD.Connect.Protocol.Sigs.eSigType.Serial, sig.Number, sig.Name, 0), sig.SigIoMask.ToIcd());
 		}
 	}
 }