+++ ICD.Connect.Telemetry.CrestronPro/Devices/AbstractFusionSigCollectionAdapter.cs
@@ -1,4 +1,5 @@
-﻿#if SIMPLSHARP
+﻿using ICD.Common.Utils.Collections;
+#if SIMPLSHARP
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -16,7 +17,7 @@ namespace ICD.Connect.Telemetry.CrestronPro.Devices
 		private readonly CrestronCollection<T> m_Collection;
 		private readonly Func<T, TAdapter> m_Factory;
 
-		private readonly Dictionary<uint, TAdapter> m_SigAdapterNumberCache;
+		private readonly IcdOrderedDictionary<uint, TAdapter> m_SigAdapterNumberCache;
 		private readonly SafeCriticalSection m_CacheSection;
 
 		#region Properties
@@ -78,7 +79,7 @@ namespace ICD.Connect.Telemetry.CrestronPro.Devices
 			m_Collection = collection;
 			m_Factory = factory;
 
-			m_SigAdapterNumberCache = new Dictionary<uint, TAdapter>();
+			m_SigAdapterNumberCache = new IcdOrderedDictionary<uint, TAdapter>();
 			m_CacheSection = new SafeCriticalSection();
 		}